feat(grpc): reduce metadata header allocations
